add Config::bloom_bits_per_key

+++ src/segment/writer/mod.rs
@@ -68,7 +68,7 @@ pub struct Writer {
 #[derive(Copy, Clone, Debug)]
 #[cfg(feature = "bloom")]
 pub enum BloomConstructionPolicy {
-    BitsPerKey(usize),
+    BitsPerKey(u8),
     FpRate(f32),
 }
